Alright that is the complete game
Now I need to add the comments, the Anubis scenes walls, floor and pillars aren't working. the walls on the space place aren't working. Apart from that everything else is working without any problems. I thought the health not going down was just a bug but looking back at it, there is no script for that so i'll implement that later. Now i need to do the comments. and explain the code.